worked on validators

+++ src/controllers/customer.controller.js
@@ -1,4 +1,7 @@
 const {customerModel} = require("../models/customer.model");
+const validator = require("../validators/customer.validator");
+const {formatZodError} = require("../utilities/errormessage");
+
 
 // get all customers
 async function getAllCustomers(req, res) {
@@ -14,6 +17,12 @@ async function getSingleCustomer(req, res) {
     //     ...req.body,
     //     id: (customers.length + 1).toString()
     // });
+    const result = validator.getCustomerValidator.safeParse(req.params.customerId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
+
     const customer = await customerModel.findById(req.params.customerId);
 
     res.json(customer).end();
@@ -25,6 +34,11 @@ async function addCustomer(req, res) {
     //     ...req.body,
     //     id: (customers.length + 1).toString()
     // });
+    const result = validator.addCustomerValidator.safeParse(req.body);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
 
     await customerModel.create({
         name: req.body.name,
@@ -45,6 +59,11 @@ async function udpateCustomer(req, res) {
 
     //     return r;
     // })
+    const result = validator.updateCustomerValidator.safeParse(req.params.customerId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
 
     await customerModel.updateOne({_id: req.params.customerId}, {...req.body});
 
@@ -54,6 +73,12 @@ async function udpateCustomer(req, res) {
 // delete customer
 async function deleteCustomer(req, res) {
     // customers = customers.filter(r => r.id !== req.params.customerId);
+    const result = validator.deleteCustomerValidator.safeParse(req.params.customerId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
+
     await customerModel.deleteOne({_id: req.params.customerId});
 
     res.json("customer deleted").end();
