+++ src/controllers/invoice.controller.js
@@ -1,4 +1,7 @@
 const {invoiceModel} = require("../models/invoice.model");
+const validator = require("../validators/invoice.validator");
+const {formatZodError} = require("../utilities/errormessage");
+
 
 // get all invoices
 async function getAllInvoice(req, res) {
@@ -14,6 +17,12 @@ async function getSingleInvoice(req, res) {
     //     ...req.body,
     //     id: (invoices.length + 1).toString()
     // });
+    const result = validator.getInvoiceValidator.safeParse(req.params.invoiceId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
+
     const invoice = await invoiceModel.findById(req.params.invoiceId);
 
     res.json(invoice).end();
@@ -25,11 +34,16 @@ async function addInvoice(req, res) {
     //     ...req.body,
     //     id: (invoices.length + 1).toString()
     // });
+    const result = validator.addInvoiceValidator.safeParse(req.body);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
 
     await invoiceModel.create({
-        name: req.body.name,
-        amount: req.body.amount,
-        location: req.body.location
+        customer_name: req.body.customer_name,
+        items: req.body.items,
+        price: req.body.price
     });
 
     res.json("invoice added").end();
@@ -44,6 +58,11 @@ async function udpateInvoice(req, res) {
 
     //     return r;
     // })
+    const result = validator.updateInvoiceValidator.safeParse(req.params.invoiceId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
 
     await invoiceModel.updateOne({_id: req.params.invoiceId}, {...req.body});
 
@@ -53,6 +72,12 @@ async function udpateInvoice(req, res) {
 // delete invoice
 async function deleteInvoice(req, res) {
     // invoices = invoices.filter(r => r.id !== req.params.invoiceId);
+    const result = validator.deleteInvoiceValidator.safeParse(req.params.invoiceId);
+
+    if (!result.success){
+        return res.status(400).json(formatZodError(result.error.issues)).end();
+    }
+
     await invoiceModel.deleteOne({_id: req.params.invoiceId});
 
     res.send("invoice deleted").end();
